Rename to kinetic scales to avoid conflict on pypi.
