+++ src/kinetic_scales/function/exponential.py
