+++ src/kinetic_scales/function/hyperbolic.py
