+++ src/kinetic_scales/function/linear.py
