+++ src/kinetic_scales/weighted/integer.py
