+++ src/kinetic_scales/weighted/rank.py
