+++ src/setup.py
@@ -6,14 +6,14 @@ from setuptools import find_packages
 
 
 setup(
-    name='scales',
+    name='kinetic-scales',
     packages=find_packages(),
-    version='',
+    version='v1.0.0',
     description='A library with some selection and weighted choice algorithms.',
     author='Real Kinetic',
     author_email='dev@realkinetic.com',
     url='https://github.com/RealKinetic/scales',
-    download_url='',
+    download_url='https://github.com/RealKinetic/scales/archive/v1.0.0.tar.gz',
     keywords=['weighted', 'random', 'choice'],
     classifiers=[],
 )
